% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Mena.R
\docType{data}
\name{Mena}
\alias{Mena}
\title{Middle East and North Africa yearly data of economic growth, inflation rate,foreign direct investment and trade}
\format{A data frame with 274 rows and 4 variables}
\usage{
data(Mena)
}
\description{
The data frame \code{Mena} contains the following variables:
\itemize{
\item{pays:}{the countries}
\item{ann:}{the year}
\item{GDPPC:}{GDP per capita growth (annual pourcentage)}
\item{FDI:}{Foreign direct investment, net inflows (pourcentageof GDP)}
\item{INF:}{Inflation, consumer prices (annual pourcentage)}
\item{OPEN:}{public capital}
}
}
\keyword{datasets}
