% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.dtp.R
\name{summary.dtp}
\alias{summary.dtp}
\title{Summary of a nardl model}
\usage{
\method{summary}{dtp}(object, ...)
}
\arguments{
\item{object}{is the object of the function}

\item{...}{not used}
}
\value{
an object of the S3 class \code{summary.dtp} with the
following components:
}
\description{
\code{summary} method for a \code{\link{dtp}} model.
}
